contributors/adobe/incoming/css3-regions/interactivity/mouse/regions-mouse-events-004.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>CSS Regions: onmouseup events work inside region</title>
        <link rel="author" title="David Alcala" href="mailto:dalcala@adobe.com">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#the-flow-into-property">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#the-flow-from-property">
        <link rel="help" href="http://www.w3.org/TR/css3-regions/#relation-to-document-events">
        <meta name="assert" content="CSS regions module does not alter the normal processing of events in the document
            tree. Therefore, onmouseup events inside a named flow should function correctly.">
        <meta name="flags" content="dom interact">
        <link href="support/css/regions-mouse-events.css" rel="stylesheet" type="text/css">
        <link href="/resources/testharness.css" rel="stylesheet" type="text/css">
        <script src="/resources/testharness.js"></script>
    </head>
    <body>
        <p id="msg">
            Mouse up on the black square.
        </p>
        <div id="content">
            <div class="block" title="FAIL"></div>
            <div class="block" title="FAIL"></div>
            <div id="target-block" class="block"></div>
        </div>
        <div class="region"></div>
        <div class="region"></div>
        <div class="region"></div>
        <br/>
        <div id="log"></div>
    </body>
     <script type="text/javascript">
        
        var testTimeout = 10000;
        setup({timeout: testTimeout});
        
        // This block is executed if running in WebKit's harness
        if (window.testRunner)
        {
        	testRunner.dumpAsText(false);
        }

        // Verify that CSS Regions are enabled in the browser.
        // Divs will be horizontal if Regions are enabled.
        // Divs will be vertical if Regions are not enabled.
        function checkLeftPosition(elemID) 
        {
            return document.getElementById(elemID).getBoundingClientRect().left;
        }
        
        if(checkLeftPosition("target-block") != 0) {

            var test1 = async_test("onmouseup event fired inside region", {timeout: testTimeout});
                test1.step( function() {
                    var testTarget = document.getElementById("target-block");
                    testTarget.onmouseup = test1.step_func( function(evt) {
                        /* Log assertion that the event got fired */
                        assert_true(true);
                        
                        /* Verify the target inside the region has the correct ID */
                        test(function() { assert_equals(evt.target.id, "target-block")}, "Target mouse upped on inside region has the correct ID");
                        
                        test1.done();
                    });
                    
                    mouseUp("target-block");
                });

            function mouseUp(block)
            {
            	if(window.testRunner)
               	{
            		var startNode = document.getElementById(block);
            		var xStartPosition = startNode.getBoundingClientRect().left;
            		var yStartPosition = startNode.getBoundingClientRect().top;
            		eventSender.mouseMoveTo(xStartPosition, yStartPosition);
            		eventSender.mouseUp();
            	}
            }
        }
        else { 
            test(function() {assert_true(false) }, "Regions are not enabled");
        }

    </script>
</html>
